+++ lib/functions/lib/client.ts
@@ -44,14 +44,6 @@ export class FunctionsInvokeClient {
       this._circuitBreaker = clientConfiguration.circuitBreaker
         ? clientConfiguration.circuitBreaker!.circuit
         : null;
-      if (!clientConfiguration.retryConfiguration) {
-        this._clientConfiguration.retryConfiguration = common.NoRetryConfigurationDetails;
-      }
-    } else {
-      // Disable default retries for the service
-      this._clientConfiguration = {
-        retryConfiguration: common.NoRetryConfigurationDetails
-      };
     }
     this._httpClient =
       params.httpClient || new common.FetchHttpClient(requestSigner, this._circuitBreaker);
@@ -161,14 +153,6 @@ export class FunctionsManagementClient {
       this._circuitBreaker = clientConfiguration.circuitBreaker
         ? clientConfiguration.circuitBreaker!.circuit
         : null;
-      if (!clientConfiguration.retryConfiguration) {
-        this._clientConfiguration.retryConfiguration = common.NoRetryConfigurationDetails;
-      }
-    } else {
-      // Disable default retries for the service
-      this._clientConfiguration = {
-        retryConfiguration: common.NoRetryConfigurationDetails
-      };
     }
     this._httpClient =
       params.httpClient || new common.FetchHttpClient(requestSigner, this._circuitBreaker);