+++ lib/managementagent/lib/model/update-management-agent-details.ts
@@ -19,7 +19,7 @@ import common = require("oci-common");
  */
 export interface UpdateManagementAgentDetails {
   /**
-   * true if the agent can be upgraded automatically; false if it must be upgraded manually. true is currently unsupported.
+   * Setting of this flag is no longer supported.
    */
   "isAgentAutoUpgradable"?: boolean;
   /**