+++ lib/ons/lib/client.ts
@@ -45,14 +45,6 @@ export class NotificationControlPlaneClient {
       this._circuitBreaker = clientConfiguration.circuitBreaker
         ? clientConfiguration.circuitBreaker!.circuit
         : null;
-      if (!clientConfiguration.retryConfiguration) {
-        this._clientConfiguration.retryConfiguration = common.NoRetryConfigurationDetails;
-      }
-    } else {
-      // Disable default retries for the service
-      this._clientConfiguration = {
-        retryConfiguration: common.NoRetryConfigurationDetails
-      };
     }
     this._httpClient =
       params.httpClient || new common.FetchHttpClient(requestSigner, this._circuitBreaker);
@@ -594,14 +586,6 @@ export class NotificationDataPlaneClient {
       this._circuitBreaker = clientConfiguration.circuitBreaker
         ? clientConfiguration.circuitBreaker!.circuit
         : null;
-      if (!clientConfiguration.retryConfiguration) {
-        this._clientConfiguration.retryConfiguration = common.NoRetryConfigurationDetails;
-      }
-    } else {
-      // Disable default retries for the service
-      this._clientConfiguration = {
-        retryConfiguration: common.NoRetryConfigurationDetails
-      };
     }
     this._httpClient =
       params.httpClient || new common.FetchHttpClient(requestSigner, this._circuitBreaker);