+++ lib/streaming/lib/client.ts
@@ -43,14 +43,6 @@ export class StreamClient {
       this._circuitBreaker = clientConfiguration.circuitBreaker
         ? clientConfiguration.circuitBreaker!.circuit
         : null;
-      if (!clientConfiguration.retryConfiguration) {
-        this._clientConfiguration.retryConfiguration = common.NoRetryConfigurationDetails;
-      }
-    } else {
-      // Disable default retries for the service
-      this._clientConfiguration = {
-        retryConfiguration: common.NoRetryConfigurationDetails
-      };
     }
     this._httpClient =
       params.httpClient || new common.FetchHttpClient(requestSigner, this._circuitBreaker);
@@ -618,14 +610,6 @@ export class StreamAdminClient {
       this._circuitBreaker = clientConfiguration.circuitBreaker
         ? clientConfiguration.circuitBreaker!.circuit
         : null;
-      if (!clientConfiguration.retryConfiguration) {
-        this._clientConfiguration.retryConfiguration = common.NoRetryConfigurationDetails;
-      }
-    } else {
-      // Disable default retries for the service
-      this._clientConfiguration = {
-        retryConfiguration: common.NoRetryConfigurationDetails
-      };
     }
     this._httpClient =
       params.httpClient || new common.FetchHttpClient(requestSigner, this._circuitBreaker);